rustls-aws-lc-rs: move SealKey to the crate root
A pure relocation from tls13.rs ahead of sharing it with the TLS1.2
message encrypters.
